Use absolute paths on the error page
For 404s, the error page is served *at* the URL that doesn't exist.
This means the correct relative paths to assets aren't fixed but depend
on the requested path.  For example, when the error page is displayed
for a path like /does/not/exist, the relative path to the stylesheet
would need to be ../../style.css.